Add files via upload
These files are from a project I completed for one of my courses in the Data Science MS program at the University of Denver. The Rmd file is a survival analysis of the hard drives run by Backblaze(www.backblaze.com), which is a company that does data cloud storage for individual and corporate clients. From this Website I downloaded hard drive data from the 4th quarter of 2019 and the first three quarters of 2020. I chose to compare the survival curves for two hard drive models. The two Python files are the programs I wrote to extract the data from the zip files I downloaded directly from the site. harddrive_data is the text file I created from the Python files so that I could conduct the survival analysis in R Studio.